@@ -8,21 +8,22 @@
     ],
     "license": "MIT",
     "require": {
-        "php": "^8.0.2",
+        "php": "^8.1",
         "calebporzio/sushi": "^2.4",
-        "doctrine/dbal": "^3.5",
+        "doctrine/dbal": "^3.0",
         "filament/filament": "^2.16",
         "filament/spatie-laravel-settings-plugin": "^2.16",
         "guzzlehttp/guzzle": "^7.2",
-        "hammerstone/fast-paginate": "^0.1.11",
-        "laravel/framework": "^9.19",
-        "laravel/sanctum": "^3.0",
+        "laravel/framework": "^10.0",
+        "laravel/sanctum": "^3.2",
         "laravel/tinker": "^2.7",
         "phpsa/filament-authentication": "^2.3",
         "saade/filament-fullcalendar": "^1.6"
     },
     "require-dev": {
+        "phpmd/phpmd": "^2.13.0",
         "barryvdh/laravel-debugbar": "^3.7",
+        "barryvdh/laravel-ide-helper": "^2.13",
         "fakerphp/faker": "^1.9.1",
         "laravel/breeze": "^1.16",
         "laravel/pint": "^1.0",
@@ -30,7 +31,7 @@
         "mockery/mockery": "^1.4.4",
         "nunomaduro/collision": "^6.1",
         "phpunit/phpunit": "^9.5.10",
-        "spatie/laravel-ignition": "^1.0"
+        "spatie/laravel-ignition": "^2.0"
     },
     "autoload": {
         "psr-4": {
