@@ -3,10 +3,16 @@
 namespace App\Filament\Resources\KategoriResource\Pages;
 
 use App\Filament\Resources\KategoriResource;
-use Filament\Pages\Actions;
 use Filament\Resources\Pages\CreateRecord;
 
 class CreateKategori extends CreateRecord
 {
+
     protected static string $resource = KategoriResource::class;
+
+    protected function getRedirectUrl() : string
+    {
+
+        return $this->getResource()::getUrl('index');
+    }
 }
