@@ -3,11 +3,16 @@
 namespace App\Filament\Resources\TimesheetResource\Pages;
 
 use App\Filament\Resources\TimesheetResource;
-use Filament\Pages\Actions;
 use Filament\Resources\Pages\CreateRecord;
 
 class CreateTimesheet extends CreateRecord
 {
+
     protected static string $resource = TimesheetResource::class;
-    
+
+    protected function getRedirectUrl() : string
+    {
+
+        return $this->getResource()::getUrl('index');
+    }
 }
