@@ -2,39 +2,28 @@
 
 namespace App\Filament\Resources;
 
-use Closure;
-use Filament\Forms;
-use Filament\Tables;
-use App\Models\Utstyr;
-use Filament\Resources\Form;
-use Filament\Resources\Table;
-use Filament\Resources\Resource;
-use Illuminate\Support\Facades\Mail;
-use Filament\Forms\Components\Select;
-use Illuminate\Support\Facades\Route;
-use Filament\Forms\Components\Repeater;
-use Filament\Tables\Actions\BulkAction;
-use Filament\Tables\Columns\TextColumn;
-use Filament\Tables\Columns\ViewColumn;
-use Illuminate\Database\Eloquent\Model;
-use Filament\Forms\Components\TextInput;
-use App\Mail\BestillUtstyr as Bestilling;
-use Filament\Tables\Columns\SelectColumn;
-use Illuminate\Database\Eloquent\Builder;
-use Filament\Tables\Actions\ReplicateAction;
-use Filament\Tables\Columns\TextInputColumn;
-use Illuminate\Database\Eloquent\Collection;
 use App\Filament\Resources\UtstyrResource\Pages;
+use App\Mail\BestillUtstyr as Bestilling;
+use App\Models\Utstyr;
+use Filament\Forms;
+use Filament\Resources\Form;
+use Filament\Resources\Resource;
+use Filament\Resources\Table;
+use Filament\Tables;
+use Illuminate\Database\Eloquent\Builder;
+use Illuminate\Database\Eloquent\Collection;
+use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\SoftDeletingScope;
-use App\Filament\Resources\UtstyrResource\RelationManagers;
+use Illuminate\Support\Facades\Mail;
 
 class UtstyrResource extends Resource
 {
+
     protected static ?string $model            = Utstyr::class;
     protected static ?string $navigationGroup  = 'Medisinsk';
     protected static ?string $modelLabel       = 'Utstyr';
     protected static ?string $pluralModelLabel = 'Utstyr';
-    protected static ?int $navigationSort      = 4;
+    protected static ?int $navigationSort      = 1;
     protected static ?string $navigationIcon   = 'heroicon-o-collection';
 
     public static function getGloballySearchableAttributes(): array
@@ -56,27 +45,31 @@ class UtstyrResource extends Resource
     {
         return $form
             ->schema([
-                TextInput::make('hva'),
-                TextInput::make('navn'),
-                TextInput::make('artikkelnummer'),
-                TextInput::make('link'),
-                Select::make('kategori')
-                    ->relationship('kategori', 'kategori'),
+                Forms\Components\TextInput::make('hva'),
+                Forms\Components\TextInput::make('navn'),
+                Forms\Components\TextInput::make('artikkelnummer'),
+                Forms\Components\TextInput::make('link'),
+                Forms\Components\Select::make('kategori')
+                                       ->relationship('kategori', 'kategori'),
             ]);
     }
 
-    public static function table(Table $table): Table
+    /**
+     * @throws \Exception
+     */
+    public static function table(Table $table) : Table
     {
+
         return $table
             ->columns([
-                TextInputColumn::make('antall')->type('number')->extraAttributes(['style' => 'width:60px']),
-                TextColumn::make('hva')->sortable(),
-                TextColumn::make('navn')->sortable(),
-                TextColumn::make('kategori.kategori')->sortable(),
-                TextColumn::make('artikkelnummer'),
-                TextColumn::make('link')
-                    ->formatStateUsing(fn () => 'Se her')
-                    ->url(fn ($record) => $record->link, true),
+                Tables\Columns\TextInputColumn::make('antall')->type('number')->extraAttributes(['style' => 'width:60px']),
+                Tables\Columns\TextColumn::make('hva')->sortable(),
+                Tables\Columns\TextColumn::make('navn')->sortable(),
+                Tables\Columns\TextColumn::make('kategori.kategori')->sortable(),
+                Tables\Columns\TextColumn::make('artikkelnummer'),
+                Tables\Columns\TextColumn::make('link')
+                                         ->formatStateUsing(fn() => 'Se her')
+                                         ->url(fn($record) => $record->link, true),
             ])
             ->filters([
                 Tables\Filters\TrashedFilter::make(),
@@ -96,19 +89,21 @@ class UtstyrResource extends Resource
                 Tables\Actions\DeleteBulkAction::make(),
                 Tables\Actions\ForceDeleteBulkAction::make(),
                 Tables\Actions\RestoreBulkAction::make(),
-                BulkAction::make('bestillValgteProdukter')
-                    ->action(function (Collection $records, array $data) {
-                        Mail::to('svinesundparken@dittapotek.no')->send(new Bestilling($records, $data));
-                    })
-                    ->form([
-                        Forms\Components\Textarea::make('info')
-                            ->label('Annen informasjon'),
-                    ])
-                    ->requiresConfirmation()
-                    ->modalHeading('Bestill utstyr')
-                    ->modalSubheading('Sikker på at du har valgt alt du trenger?')
-                    ->modalButton('ja, bestill utstyr!')
-                    ->deselectRecordsAfterCompletion()
+                Tables\Actions\BulkAction::make('bestillValgteProdukter')
+                                         ->action(function (Collection $records, array $data) {
+
+                                             Mail::to('svinesundparken@dittapotek.no')->send(new Bestilling($records, $data));
+                                         })
+                                         ->form([
+                                             Forms\Components\Textarea::make('info')
+                                                                      ->label('Annen informasjon'),
+                                         ])
+                                         ->requiresConfirmation()
+                                         ->modalHeading('Bestill utstyr')
+                                         ->modalSubheading('Sikker på at du har valgt alt du trenger?')
+                                         ->modalContent(fn($records) => view('filament.pages.modalUtstyr', ['records' => $records]))
+                                         ->modalButton('ja, bestill utstyr!')
+                                         ->deselectRecordsAfterCompletion()->modalWidth('lg')
             ]);
     }
 
