@@ -2,15 +2,13 @@
 
 namespace App\Models;
 
-use Cache;
-use Carbon\Carbon;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\SoftDeletes;
-use Illuminate\Database\Eloquent\Builder;
-use App\Notifications\UserUnavailable;
 use Illuminate\Notifications\Notifiable;
-use Illuminate\Notifications\DatabaseNotification;
 
+/**
+ * @mixin IdeHelperEconomy
+ */
 class Economy extends Model
 {
     use Notifiable;
